adding some logic to scripts

+++ codedeploy/web/scripts/packages.sh
@@ -1,4 +1,13 @@
 #!/bin/bash
 apt-get install -y nginx-full uwsgi uwsgi-plugin-python
-rm /etc/nginx/sites-available/default
-mkdir /data
+
+#if file exists, remove it
+if [ -f "/etc/nginx/sites-available/default"]
+then
+  rm /etc/nginx/sites-available/default
+fi
+#if /data directory does not exist, create it
+if [ ! -d "/data"]
+then
+  mkdir /data
+fi
